Refactored AbstractQuest and Questbasic to just Quest
